requirements.txt
aioice==0.9.0
aiortc==1.11.0
annotated-types==0.7.0
anyio==4.9.0
av==14.3.0
bidict==0.23.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
fastapi==0.115.12
google-crc32c==1.7.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
idna==3.10
ifaddr==0.2.0
numpy==2.2.4
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.0.0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
scipy==1.15.2
setuptools==78.1.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
webrtcvad==2.0.10
wsproto==1.2.0
